+++ packages/components/organisms/f-checkout/src/store/checkout.module.js
@@ -7,7 +7,8 @@ import {
     UPDATE_CUSTOMER_DETAILS,
     UPDATE_FULFILMENT_ADDRESS,
     UPDATE_FULFILMENT_TIME,
-    UPDATE_STATE
+    UPDATE_STATE,
+    UPDATE_USER_NOTE
 } from './mutation-types';
 
 export default {
@@ -32,7 +33,7 @@ export default {
             city: '',
             postcode: ''
         },
-        notes: [],
+        userNote: '',
         isFulfillable: true,
         notices: [],
         messages: [],
@@ -48,7 +49,7 @@ export default {
         /**
          * Get the checkout details from the backend and update the state.
          *
-         * @param {Object} commit - Automatically handled by Vuex to be able to commit mutations.
+         * @param {Object} context - Vuex context object, this is the standard first parameter for actions.
          * @param {Object} payload - Parameter with the different configurations for the request.
          */
         getCheckout: async ({ commit, state }, { url, timeout }) => {
@@ -73,12 +74,11 @@ export default {
         /**
          * Post the checkout details to the backend.
          *
-         * @param {Object} commit - Automatically handled by Vuex to be able to commit mutations.
-         * @param {Object} state - Automatically handled by Vuex to be able to retrieve state.
+         * @param {Object} context - Vuex context object, this is the standard first parameter for actions.
          * @param {Object} payload - Parameter with the different configurations for the request.
          */
         // eslint-disable-next-line no-unused-vars
-        patchCheckout: async ({ commit, state }, {
+        patchCheckout: async ({ state }, {
             url, data, timeout
         }) => {
             // TODO: deal with exceptions and handle this action properly (when the functionality is ready)
@@ -101,12 +101,10 @@ export default {
         /**
          * Post the guest user details to the backend.
          *
-         * @param {Object} commit - Automatically handled by Vuex to be able to commit mutations.
-         * @param {Object} state - Automatically handled by Vuex to be able to retrieve state.
+         * @param {Object} context - Vuex context object, this is the standard first parameter for actions
          * @param {Object} payload - Parameter with the different configurations for the request.
          */
-        // eslint-disable-next-line no-unused-vars
-        createGuestUser: async ({ commit, state }, {
+        createGuestUser: async (context, {
             url, tenant, data, timeout
         }) => {
             const config = {
@@ -126,7 +124,7 @@ export default {
         /**
          * Get the fulfilment details from the backend and update the state.
          *
-         * @param {Object} commit - Automatically handled by Vuex to be able to commit mutations.
+         * @param {Object} context - Vuex context object, this is the standard first parameter for actions
          * @param {Object} payload - Parameter with the different configurations for the request.
          */
         getAvailableFulfilment: async ({ commit }, { url, timeout }) => {
@@ -146,7 +144,7 @@ export default {
         /**
          * Get the basket details from the backend and update the state.
          *
-         * @param {Object} commit - Automatically handled by Vuex to be able to commit mutations.
+         * @param {Object} context - Vuex context object, this is the standard first parameter for actions
          * @param {Object} payload - Parameter with the different configurations for the request.
          */
         getBasket: async ({ commit }, {
@@ -187,6 +185,10 @@ export default {
 
         updateFulfilmentTime ({ commit }, payload) {
             commit(UPDATE_FULFILMENT_TIME, payload);
+        },
+
+        updateUserNote ({ commit }, payload) {
+            commit(UPDATE_USER_NOTE, payload);
         }
     },
 
@@ -197,7 +199,6 @@ export default {
             customer,
             address,
             time,
-            notes,
             isFulfillable,
             notices,
             messages
@@ -222,7 +223,6 @@ export default {
                 state.address.postcode = address.postalCode;
             }
 
-            state.notes = notes;
             state.isFulfillable = isFulfillable;
             state.notices = notices;
             state.messages = messages;
@@ -264,6 +264,10 @@ export default {
                 ...state.time,
                 ...time
             };
+        },
+
+        [UPDATE_USER_NOTE]: (state, userNote) => {
+            state.userNote = userNote;
         }
     }
 };