+++ packages/components/organisms/f-checkout/test/specs/component/f-checkout.component.spec.js
@@ -2,8 +2,7 @@ import CheckoutComponent from '../../../test-utils/component-objects/f-checkout.
 
 describe('f-checkout component tests', () => {
     before(() => {
-        browser.url('?path=/story/components-organisms--checkout-component');
-        browser.switchToFrame(0);
+        browser.url('iframe.html?id=components-organisms--checkout-component&knob-Checkout%20Url=%2Fcheckout-delivery.json&knob-Available%20Fulfilment%20Url=%2Fcheckout-available-fulfilment.json&knob-Create%20Guest%20Url=%2Fcreate-guest.json&knob-Auth%20token=a&knob-Login%20Url=%2Flogin&viewMode=story');
         CheckoutComponent.waitForCheckoutComponent();
     });
 
@@ -30,6 +29,11 @@ describe('f-checkout component tests', () => {
 
         // Assert
         // Waiting for route here, so we can grab redirect url and show form submits.
+    });
+
+    it('should display the mandatory "mobileNumber" field', () => {
+        // Assert
+        expect(CheckoutComponent.isFieldDisplayed('mobileNumber')).toBe(true);
     });
 
     it('should display a "mobileNumber" error message when an unsupported country code is used in the mobile number field', () => {